37_resnet/requirements.txt
# ResNet Implementation Requirements

# Core dependencies
torch>=1.9.0
torchvision>=0.10.0

# Data handling
numpy>=1.20.0
pillow>=8.0.0

# Configuration and utilities  
pyyaml>=5.4.0

# Visualization (optional)
matplotlib>=3.3.0
seaborn>=0.11.0

# Progress tracking (optional)
tqdm>=4.60.0

# Development (optional)
jupyter>=1.0.0
tensorboard>=2.7.0